Fix/修改relation服务 (#42)

+++ biz/application/service/relation.go
@@ -2,12 +2,20 @@ package service
 
 import (
 	"context"
+	"github.com/CloudStriver/go-pkg/utils/pagination"
 	"github.com/CloudStriver/go-pkg/utils/pconvertor"
+	"github.com/CloudStriver/go-pkg/utils/util/log"
 	"github.com/CloudStriver/platform/biz/infrastructure/config"
+	"github.com/CloudStriver/platform/biz/infrastructure/consts"
+	"github.com/CloudStriver/platform/biz/infrastructure/convertor"
 	relationmapper "github.com/CloudStriver/platform/biz/infrastructure/mapper/relation"
+	"github.com/CloudStriver/platform/biz/infrastructure/sort"
 	"github.com/CloudStriver/service-idl-gen-go/kitex_gen/platform"
 	"github.com/google/wire"
+	"github.com/samber/lo"
 	"github.com/zeromicro/go-zero/core/stores/redis"
+	"go.mongodb.org/mongo-driver/bson/primitive"
+	"go.mongodb.org/mongo-driver/mongo"
 )
 
 type RelationService interface {
@@ -27,16 +35,21 @@ var RelationSet = wire.NewSet(
 )
 
 type RelationServiceImpl struct {
-	Config        *config.Config
-	Redis         *redis.Redis
-	RelationModel relationmapper.RelationNeo4jMapper
+	Config              *config.Config
+	Redis               *redis.Redis
+	RelationModel       relationmapper.RelationNeo4jMapper
+	RelationMongoMapper relationmapper.IMongoMapper
 }
 
 func (s *RelationServiceImpl) GetRelationPathsCount(ctx context.Context, req *platform.GetRelationPathsCountReq) (resp *platform.GetRelationPathsCountResp, err error) {
 	resp = new(platform.GetRelationPathsCountResp)
-	resp.Total, err = s.RelationModel.GetRelationPathsCount(ctx, req.FromType1, req.FromId1, req.FromType2, req.EdgeType1, req.EdgeType2, req.ToType)
-	if err != nil {
-		return resp, err
+	if s.Config.Neo4jConf.Enable {
+		resp.Total, err = s.RelationModel.GetRelationPathsCount(ctx, req.FromType1, req.FromId1, req.FromType2, req.EdgeType1, req.EdgeType2, req.ToType)
+		if err != nil {
+			return resp, err
+		}
+	} else {
+		return resp, consts.ErrComponentNotStarted
 	}
 	return resp, nil
 
@@ -44,18 +57,60 @@ func (s *RelationServiceImpl) GetRelationPathsCount(ctx context.Context, req *pl
 
 func (s *RelationServiceImpl) DeleteNode(ctx context.Context, req *platform.DeleteNodeReq) (resp *platform.DeleteNodeResp, err error) {
 	resp = new(platform.DeleteNodeResp)
-	if err = s.RelationModel.DeleteNode(ctx, req.NodeId, req.NodeType); err != nil {
+
+	tx := s.RelationMongoMapper.StartClient()
+	if err = tx.UseSession(ctx, func(sessionContext mongo.SessionContext) error {
+		var err1 error
+		if err1 = sessionContext.StartTransaction(); err1 != nil {
+			return err1
+		}
+		if _, err1 = s.RelationMongoMapper.Delete(sessionContext, &relationmapper.FilterOptions{
+			OnlyFromType: lo.ToPtr(req.NodeType),
+			OnlyFromId:   lo.ToPtr(req.NodeId),
+		}); err1 != nil {
+			if rbErr := sessionContext.AbortTransaction(sessionContext); rbErr != nil {
+				log.CtxError(sessionContext, "保存文件中产生错误[%v]: 回滚异常[%v]\n", err1, rbErr)
+			}
+			return err1
+		}
+
+		if _, err1 = s.RelationMongoMapper.Delete(sessionContext, &relationmapper.FilterOptions{
+			OnlyToType: lo.ToPtr(req.NodeType),
+			OnlyToId:   lo.ToPtr(req.NodeId),
+		}); err1 != nil {
+			if rbErr := sessionContext.AbortTransaction(sessionContext); rbErr != nil {
+				log.CtxError(sessionContext, "保存文件中产生错误[%v]: 回滚异常[%v]\n", err1, rbErr)
+			}
+			return err1
+		}
+
+		if err1 = sessionContext.CommitTransaction(sessionContext); err1 != nil {
+			log.CtxError(sessionContext, "保存文件: 提交事务异常[%v]\n", err1)
+			return err1
+		}
+		return nil
+	}); err != nil {
 		return resp, err
 	}
+
+	if s.Config.Neo4jConf.Enable {
+		if err = s.RelationModel.DeleteNode(ctx, req.NodeId, req.NodeType); err != nil {
+			return resp, err
+		}
+	}
+
 	return resp, nil
 }
 
 func (s *RelationServiceImpl) GetRelationPaths(ctx context.Context, req *platform.GetRelationPathsReq) (resp *platform.GetRelationPathsResp, err error) {
 	resp = new(platform.GetRelationPathsResp)
-	p := pconvertor.PaginationOptionsToModelPaginationOptions(req.PaginationOptions)
-	resp.Relations, err = s.RelationModel.GetRelationPaths(ctx, req.FromType, req.FromId, req.EdgeType1, req.EdgeType2, p)
-	if err != nil {
-		return resp, err
+	if s.Config.Neo4jConf.Enable {
+		p := pconvertor.PaginationOptionsToModelPaginationOptions(req.PaginationOptions)
+		if resp.Relations, err = s.RelationModel.GetRelationPaths(ctx, req.FromType, req.FromId, req.EdgeType1, req.EdgeType2, p); err != nil {
+			return resp, err
+		}
+	} else {
+		return resp, consts.ErrComponentNotStarted
 	}
 	return resp, nil
 }
@@ -64,9 +119,23 @@ func (s *RelationServiceImpl) GetRelationCount(ctx context.Context, req *platfor
 	resp = new(platform.GetRelationCountResp)
 	switch o := req.RelationFilterOptions.(type) {
 	case *platform.GetRelationCountReq_FromFilterOptions:
-		resp.Total, err = s.RelationModel.MatchFromEdgesCount(ctx, o.FromFilterOptions.FromType, o.FromFilterOptions.FromId, o.FromFilterOptions.ToType, req.RelationType)
+		if _, resp.Total, err = s.RelationMongoMapper.FindManyAndCount(ctx, &relationmapper.FilterOptions{
+			OnlyFromType:     lo.ToPtr(o.FromFilterOptions.FromType),
+			OnlyFromId:       lo.ToPtr(o.FromFilterOptions.FromId),
+			OnlyToType:       lo.ToPtr(o.FromFilterOptions.ToType),
+			OnlyRelationType: lo.ToPtr(req.RelationType),
+		}, &pagination.PaginationOptions{}, sort.TimeCursorType); err != nil {
+			return resp, err
+		}
 	case *platform.GetRelationCountReq_ToFilterOptions:
-		resp.Total, err = s.RelationModel.MatchToEdgesCount(ctx, o.ToFilterOptions.ToType, o.ToFilterOptions.ToId, o.ToFilterOptions.FromType, req.RelationType)
+		if _, resp.Total, err = s.RelationMongoMapper.FindManyAndCount(ctx, &relationmapper.FilterOptions{
+			OnlyFromType:     lo.ToPtr(o.ToFilterOptions.FromType),
+			OnlyToId:         lo.ToPtr(o.ToFilterOptions.ToId),
+			OnlyToType:       lo.ToPtr(o.ToFilterOptions.ToType),
+			OnlyRelationType: lo.ToPtr(req.RelationType),
+		}, &pagination.PaginationOptions{}, sort.TimeCursorType); err != nil {
+			return resp, err
+		}
 	}
 	if err != nil {
 		return resp, err
@@ -76,14 +145,48 @@ func (s *RelationServiceImpl) GetRelationCount(ctx context.Context, req *platfor
 
 func (s *RelationServiceImpl) GetRelations(ctx context.Context, req *platform.GetRelationsReq) (resp *platform.GetRelationsResp, err error) {
 	resp = new(platform.GetRelationsResp)
+
+	var (
+		total     int64
+		relations []*relationmapper.Relation
+	)
+
 	p := pconvertor.PaginationOptionsToModelPaginationOptions(req.PaginationOptions)
 	switch o := req.RelationFilterOptions.(type) {
 	case *platform.GetRelationsReq_FromFilterOptions:
-		resp.Relations, resp.Total, err = s.RelationModel.MatchFromEdgesAndCount(ctx, o.FromFilterOptions.FromType, o.FromFilterOptions.FromId, o.FromFilterOptions.ToType,
-			req.RelationType, p)
+		if relations, total, err = s.RelationMongoMapper.FindManyAndCount(ctx, &relationmapper.FilterOptions{
+			OnlyFromType:     lo.ToPtr(o.FromFilterOptions.FromType),
+			OnlyFromId:       lo.ToPtr(o.FromFilterOptions.FromId),
+			OnlyToType:       lo.ToPtr(o.FromFilterOptions.ToType),
+			OnlyRelationType: lo.ToPtr(req.RelationType),
+		}, p, sort.TimeCursorType); err != nil {
+			return resp, err
+		}
+
+		if p.LastToken != nil {
+			resp.Token = *p.LastToken
+		}
+		resp.Total = total
+		resp.Relations = lo.Map(relations, func(comment *relationmapper.Relation, _ int) *platform.Relation {
+			return convertor.RelationMapperToRelation(comment)
+		})
 	case *platform.GetRelationsReq_ToFilterOptions:
-		resp.Relations, resp.Total, err = s.RelationModel.MatchToEdgesAndCount(ctx, o.ToFilterOptions.ToType, o.ToFilterOptions.ToId, o.ToFilterOptions.FromType,
-			req.RelationType, p)
+		if relations, total, err = s.RelationMongoMapper.FindManyAndCount(ctx, &relationmapper.FilterOptions{
+			OnlyFromType:     lo.ToPtr(o.ToFilterOptions.FromType),
+			OnlyToId:         lo.ToPtr(o.ToFilterOptions.ToId),
+			OnlyToType:       lo.ToPtr(o.ToFilterOptions.ToType),
+			OnlyRelationType: lo.ToPtr(req.RelationType),
+		}, p, sort.TimeCursorType); err != nil {
+			return resp, err
+		}
+
+		if p.LastToken != nil {
+			resp.Token = *p.LastToken
+		}
+		resp.Total = total
+		resp.Relations = lo.Map(relations, func(comment *relationmapper.Relation, _ int) *platform.Relation {
+			return convertor.RelationMapperToRelation(comment)
+		})
 	}
 	if err != nil {
 		return resp, err
@@ -92,32 +195,20 @@ func (s *RelationServiceImpl) GetRelations(ctx context.Context, req *platform.Ge
 }
 
 func (s *RelationServiceImpl) DeleteRelation(ctx context.Context, req *platform.DeleteRelationReq) (resp *platform.DeleteRelationResp, err error) {
-	if err = s.RelationModel.DeleteEdge(ctx, &platform.Relation{
-		FromType:     req.FromType,
-		FromId:       req.FromId,
-		ToType:       req.ToType,
-		ToId:         req.ToId,
-		RelationType: req.RelationType,
+	resp = new(platform.DeleteRelationResp)
+
+	if _, err = s.RelationMongoMapper.Delete(ctx, &relationmapper.FilterOptions{
+		OnlyFromType:     lo.ToPtr(req.FromType),
+		OnlyFromId:       lo.ToPtr(req.FromId),
+		OnlyToType:       lo.ToPtr(req.ToType),
+		OnlyToId:         lo.ToPtr(req.ToId),
+		OnlyRelationType: lo.ToPtr(req.RelationType),
 	}); err != nil {
 		return resp, err
 	}
-	return resp, nil
-}
 
-func (s *RelationServiceImpl) CreateRelation(ctx context.Context, req *platform.CreateRelationReq) (resp *platform.CreateRelationResp, err error) {
-	resp = new(platform.CreateRelationResp)
-	ok, err := s.RelationModel.MatchEdge(ctx, &platform.Relation{
-		FromType:     req.FromType,
-		FromId:       req.FromId,
-		ToType:       req.ToType,
-		ToId:         req.ToId,
-		RelationType: req.RelationType,
-	})
-	if err != nil {
-		return resp, err
-	}
-	if !ok {
-		if err = s.RelationModel.CreateEdge(ctx, &platform.Relation{
+	if s.Config.Neo4jConf.Enable {
+		if err = s.RelationModel.DeleteEdge(ctx, &platform.Relation{
 			FromType:     req.FromType,
 			FromId:       req.FromId,
 			ToType:       req.ToType,
@@ -126,14 +217,16 @@ func (s *RelationServiceImpl) CreateRelation(ctx context.Context, req *platform.
 		}); err != nil {
 			return resp, err
 		}
-		resp.Ok = true
 	}
+
 	return resp, nil
 }
 
-func (s *RelationServiceImpl) GetRelation(ctx context.Context, req *platform.GetRelationReq) (resp *platform.GetRelationResp, err error) {
-	resp = new(platform.GetRelationResp)
-	if resp.Ok, err = s.RelationModel.MatchEdge(ctx, &platform.Relation{
+func (s *RelationServiceImpl) CreateRelation(ctx context.Context, req *platform.CreateRelationReq) (resp *platform.CreateRelationResp, err error) {
+	resp = new(platform.CreateRelationResp)
+
+	var res *platform.GetRelationResp
+	if res, err = s.GetRelation(ctx, &platform.GetRelationReq{
 		FromType:     req.FromType,
 		FromId:       req.FromId,
 		ToType:       req.ToType,
@@ -142,5 +235,52 @@ func (s *RelationServiceImpl) GetRelation(ctx context.Context, req *platform.Get
 	}); err != nil {
 		return resp, err
 	}
+
+	if !res.Ok {
+		if _, err = s.RelationMongoMapper.Insert(ctx, &relationmapper.Relation{
+			ID:           primitive.NilObjectID,
+			FromType:     req.FromType,
+			FromId:       req.FromId,
+			ToType:       req.ToType,
+			ToId:         req.ToId,
+			RelationType: req.RelationType,
+		}); err != nil {
+			return resp, err
+		}
+
+		if s.Config.Neo4jConf.Enable {
+			if err = s.RelationModel.CreateEdge(ctx, &platform.Relation{
+				FromType:     req.FromType,
+				FromId:       req.FromId,
+				ToType:       req.ToType,
+				ToId:         req.ToId,
+				RelationType: req.RelationType,
+			}); err != nil {
+				return resp, err
+			}
+		}
+		resp.Ok = true
+	}
+	return resp, nil
+}
+
+func (s *RelationServiceImpl) GetRelation(ctx context.Context, req *platform.GetRelationReq) (resp *platform.GetRelationResp, err error) {
+	resp = new(platform.GetRelationResp)
+
+	var relation *relationmapper.Relation
+	if relation, err = s.RelationMongoMapper.FindOne(ctx, &relationmapper.FilterOptions{
+		OnlyFromType:     lo.ToPtr(req.FromType),
+		OnlyFromId:       lo.ToPtr(req.FromId),
+		OnlyToType:       lo.ToPtr(req.ToType),
+		OnlyToId:         lo.ToPtr(req.ToId),
+		OnlyRelationType: lo.ToPtr(req.RelationType),
+	}); err != nil {
+		return resp, err
+	}
+
+	if relation != nil {
+		resp.Ok = true
+	}
+
 	return resp, nil
 }
